+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.27.0 (April 12, 2016)
+- Added `context.documents.addMember()` method, which lets you add a single member reference to a Container
+- Added `context.documents.addMember()` method, same as `persistedContainer.addMember()`
+- Added `context.documents.addMembers()` method, which lets you add a multiple members references to a Container
+- Added `context.documents.addMembers()` method, same as `persistedContainer.addMembers()`
+
+#### Breaking Changes
+- Changed `context.documents.delete()` method, now it accepts a document URI and not a PersistedDocument object
+
+# 0.26.0 (April 12, 2016)
+- Added `persistedContainer.getMembers()` method, which lets you obtain the members of the Container
+
+#### Breaking Changes
+- Renamed `app.allowsOrigin` property to `app.allowsOrigins`, and now it is an array of `Pointer.Class` or `string`.
+
 # 0.25.1 (April 12, 2016)
 - `Document.Factory.createFrom()` now accepts a object with nested object properties and structures it as a Document with fragments
 - `context.documents.createChild()` same ^
@@ -12,7 +27,7 @@
 - Added `LDP/IndirectContainer` module
 - Added `context.documents.createAccessPoint()` method, which lets you create an AccessPoint for a document
 - Added `context.documents.createAccessPoint()` method, same as `persistedDocument.createAccessPoint()`
-- `context.documents.createChile()` now accepts relative URIs
+- `context.documents.createChild()` now accepts relative URIs
 - `jsonLDConverter` now expands pointers even if they are out of the context's scope
 
 # 0.23.2 (April 8, 2016)